+++ src/extension.ts
@@ -126,7 +126,7 @@ export async function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(
     vscode.commands.registerCommand(
       'verilog.instantiateModule',
-      ModuleInstantiation.instantiateModuleInteract
+      () => ModuleInstantiation.instantiateModuleInteract(ctagsManager)
     )
   );